Fixing nearly all of the MS Visual Studio warnings
At least an attempt as I don't have that compiler available here.

I have fixed everything up to
 ``` ..\engine\src\Enemy.cpp(60): warning C4355: 'this' : used in base member initializer list```
as I don't know how to fix that compiler warning.

Also I left out all of UtilsMath.hThings I did not fix:

Reference #994

+++ src/GameStateCutscene.cpp
@@ -93,7 +93,7 @@ bool Scene::logic(FPoint *caption_margins, bool scale_graphics) {
 
 		if (components.front().type == "caption") {
 
-			int caption_width = render_device->getContextSize().w - (render_device->getContextSize().w * (caption_margins->x * 2.0f));
+			int caption_width = render_device->getContextSize().w - (int)(render_device->getContextSize().w * (caption_margins->x * 2.0f));
 			font->setFont("font_captions");
 			caption = components.front().s;
 			caption_size = font->calc_size(caption, caption_width);
